requirements.txt
requests
Django
Pillow
django_imagekit
djangorestframework
django-celery

